+++ test/sparkledriver/core_test.clj
@@ -109,12 +109,30 @@
       (is (= (title browser) "The Second Coming")))
 
     (testing "cookies"
-      (add-cookie browser "hello" "SparkleDriver" {:domain "example.com" :path "/home"})
-      (is (= (browser-cookies->map browser)
-             {:hello
-              {:value "SparkleDriver" :domain "example.com" :path "/home"}}))
-      (delete-all-cookies browser)
-      (is (= (browser-cookies->map browser) {})))))
+      (is (= (-> browser
+                 (set-cookie! "no-domain" "uses current domain")
+                 (set-cookie! "no-domain" "can be changed")
+                 (set-cookie! "with-domain" "SparkleDriver" :domain "example.org" :path "/home")
+                 browser-cookies->map)
+             {:no-domain   {:domain "0.0.0.0", :path "/", :value "can be changed"},
+              :with-domain {:domain "example.org", :path "/home", :value "SparkleDriver"}}))
+
+      (is (= (-> browser
+                 delete-all-cookies!
+                 browser-cookies->map) {}))
+
+      (is (= (-> browser
+                 (set-cookie! "a-cookie" "no domain" :path "/foo")
+                 (set-cookie! "a-cookie" "no domain" :path "/bar")
+                 (set-cookie! "a-cookie" "some domain" :domain "example.org")
+                 (set-cookie! "a-cookie" "some domain" :domain "example.org" :http-only true, :secure true, :path "/foo")
+                 (set-cookie! "a-cookie" "some other domain" :domain "clojure.org")
+                 (.. manage getCookies)
+                 (->> (map str))
+                 sort)
+             ["a-cookie=no domain; path=/bar; domain=0.0.0.0"
+              "a-cookie=some domain; path=/foo; domain=example.org;secure;"
+              "a-cookie=some other domain; path=/; domain=clojure.org"])))))
 
 (deftest fetch-test
   (with-browser [browser (make-browser)]